fix: 만든 아이콘 컴포넌트들 경로 수정 및 자잘한 버그 수정
- viewBox에서 단위 제거
- clip-path 제거
- file-rule, clip-rule 제거
