--- src/components/common/@Icons/New/LogoWhite.tsx
@@ -1,19 +1,0 @@
-import { Icon, IconProps } from '@chakra-ui/react';
-
-const LogoWhiteIcon = ({ ...props }: IconProps) => {
-  const w = props.w? props.w : "120px"
-  const h = props.h? props.h : "16px"
-  return (
-    <Icon
-      width={w} height={h} viewBox="0 0 120px 16px" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
-<path d="M0 0H13.1532V35.989H0.242621L0 0Z" fill="white"/>
-<path d="M16.4419 0H27.3683L41.5989 19.9427V0H52.6133V35.989H41.5989L27.4564 16.2558V35.989H16.4419V0Z" fill="white"/>
-<path fill-rule="evenodd" clip-rule="evenodd" d="M241.431 21.3699C247.573 19.8365 253.887 19.0601 260.226 19.0588C260.959 19.0313 261.549 18.7153 262 18.2421V35.6705H251.137C253.871 34.7206 256.565 33.6579 259.212 32.485C259.657 32.2842 260.029 31.9552 260.278 31.5421C260.527 31.129 260.64 30.6516 260.602 30.1739C260.602 28.4498 259.061 25.7719 257.069 25.7719C256.886 25.7131 256.688 25.7131 256.505 25.7719C251.768 28.1239 246.701 29.7802 241.469 30.6875C241.44 29.8687 241.396 29.0605 241.351 28.2548C241.278 26.9307 241.205 25.6135 241.205 24.2679C241.168 23.2967 241.244 22.3245 241.431 21.3699ZM235.105 35.6705H215.491C216.954 34.887 218.352 33.984 219.671 32.9686C223.544 29.9858 226.611 26.1232 228.602 21.7179C229.22 20.3472 229.529 18.8633 229.509 17.3661C229.509 15.1533 228.64 13.3461 226.108 13.3461C225.749 13.3461 225.39 13.3923 225.031 13.4384C224.672 13.4845 224.313 13.5306 223.954 13.5306C219.91 13.8994 215.867 14.2313 211.823 14.2313C209.127 14.2257 206.436 14.016 203.774 13.6043C205.41 11.6733 207.42 10.0767 209.69 8.9059C211.96 7.73508 214.444 7.01309 217.001 6.78156H220.213C222.291 6.78156 224.521 6.5234 224.521 4.01559C224.528 2.8852 224.165 1.78218 223.486 0.866838C223.249 0.547903 222.978 0.257563 222.679 0L251.765 0C251.39 0.000615993 251.016 0.00730619 250.641 0.0200812C246.167 0.0200812 241.619 0.276915 237.146 0.643749C235.341 0.790483 234.589 3.72516 234.589 5.11913C234.529 6.18737 234.862 7.24149 235.529 8.09048V13.8498C235.002 13.8472 234.481 13.9527 233.998 14.1596C233.516 14.3664 233.083 14.67 232.729 15.0505C232.374 15.431 232.106 15.8799 231.941 16.3682C231.775 16.8565 231.717 17.3734 231.77 17.885C231.741 18.7132 232 19.5265 232.505 20.1926C233.01 20.8587 233.732 21.3384 234.552 21.5534C234.42 23.9762 234.558 26.4115 234.695 28.8375C234.793 30.562 234.89 32.2818 234.89 33.989C234.914 34.5818 234.982 35.1472 235.105 35.6705ZM251.879 0C253.82 0.00319072 255.757 0.169345 257.67 0.496984C259.212 0.863818 259.588 2.91809 259.588 4.16533C259.588 5.41257 258.911 8.16381 257.52 8.16381C255.791 8.16381 254.08 7.94372 252.37 7.72363C250.66 7.50354 248.949 7.28344 247.22 7.28344C245.541 7.26492 243.88 7.62922 242.371 8.3473C242.174 8.8892 242.072 9.4597 242.07 10.0347C242.146 11.0727 242.146 12.1148 242.07 13.1528C243.555 13.0447 245.033 12.9142 246.509 12.784C250.039 12.4725 253.554 12.1623 257.106 12.1623C257.689 12.1623 258.262 12.0706 258.836 11.9789C259.409 11.8872 259.982 11.7955 260.565 11.7955C261.163 11.7955 261.634 11.9535 262 12.221V0L251.879 0ZM191 35.6705H200.692C200.301 34.9122 200.07 34.1295 200.07 33.4456C200.06 33.0714 200.127 32.6989 200.267 32.3503C200.406 32.0017 200.616 31.6839 200.884 31.4158C201.152 31.1477 201.472 30.9346 201.825 30.7891C202.179 30.6437 202.559 30.5688 202.942 30.569C207.911 30.3573 212.706 28.729 216.736 25.8853C218.188 24.8051 219.258 23.3095 219.797 21.6073H215.338C211.181 21.6073 206.986 21.9023 202.829 22.2711H201.431C200.146 22.2711 196.178 21.976 195.535 20.6483C195.035 19.5449 194.778 18.3514 194.779 17.1448C195.071 15.297 195.735 13.5238 196.734 11.9285C197.733 10.3331 199.047 8.94723 200.599 7.8511C203.868 4.13213 208.125 1.42838 212.846 0H160.02L159.777 0.239185C159.777 7.68151 159.55 15.1239 159.285 22.8267C159.272 26.3806 158.067 29.8311 155.857 32.6432C154.955 33.7922 153.905 34.8084 152.74 35.6705H165.109C165.156 34.7038 165.26 33.7377 165.364 32.7741C165.503 31.4864 165.642 30.2032 165.642 28.9289C165.642 26.7527 164.943 26.008 164.327 25.3518C163.842 24.8349 163.409 24.3729 163.409 23.3096C163.492 21.6964 164.164 20.1643 165.307 18.987C165.307 18.28 165.354 17.5822 165.401 16.8889C165.447 16.2045 165.493 15.5246 165.493 14.8446C165.487 13.0518 165.388 11.2605 165.195 9.47748C164.748 9.15178 164.389 8.72595 164.15 8.23661C163.91 7.74727 163.796 7.20901 163.818 6.66783C163.862 5.25963 164.447 3.91782 165.458 2.90511C166.47 1.8924 167.836 1.28165 169.288 1.19254H170.591C177.94 0.798927 185.293 1.90336 192.173 4.43443C192.916 4.7974 193.533 5.36215 193.95 6.05949C194.366 6.75683 194.563 7.55661 194.517 8.36083C194.517 12.6137 185.821 17.5906 181.408 20.1159C181 20.3498 180.628 20.5627 180.303 20.752C185.345 22.4856 190.004 25.1254 194.034 28.5326C194.531 29.0783 194.797 29.7861 194.778 30.5138C194.592 32.0627 193.215 35.6648 191.02 35.6648L191 35.6705ZM190.444 35.6705L190.424 35.6648C189.299 34.9775 188.205 34.259 187.114 33.542C185.861 32.7183 184.61 31.8964 183.317 31.1261C180.656 29.5469 177.897 28.1275 175.056 26.8757C174.206 26.4703 173.284 26.2258 172.34 26.1552H172.116C171.844 28.6316 171.782 31.1254 171.93 33.6116C171.949 33.9475 171.949 34.2842 171.93 34.6202C171.833 34.9833 171.7 35.3348 171.536 35.6705H190.444ZM120.738 35.6705H135.817C135.603 35.1582 135.484 34.5758 135.484 33.9158C135.358 31.9271 135.4 29.8885 135.443 27.8668V27.8664V27.8661V27.8658C135.463 26.8615 135.484 25.8613 135.484 24.8734C135.484 21.3941 135.532 17.9241 135.579 14.4541C135.626 10.9841 135.673 7.51414 135.673 4.03485C135.673 2.09984 138.133 0.611325 139.949 0.611325C140.286 0.576163 140.626 0.615598 140.946 0.726589C141.265 0.83758 141.555 1.01736 141.794 1.25283C142.034 1.4883 142.217 1.77343 142.33 2.08739C142.442 2.40135 142.482 2.73616 142.447 3.06731V3.36506C142.447 6.17454 142.267 8.9654 142.087 11.7563C141.907 14.5471 141.728 17.338 141.728 20.1474C141.626 22.6344 141.752 25.1253 142.106 27.5898C142.447 29.1154 143.128 30.4179 144.906 30.4179C145.783 30.3478 146.634 30.094 147.403 29.6736C152.89 27.0688 153.004 24.0547 153.004 18.5101C153.004 14.789 152.777 11.0678 152.55 7.34662C152.55 6.43954 152.465 5.51404 152.381 4.58387L152.38 4.5836C152.295 3.64407 152.209 2.69978 152.209 1.76488C152.209 1.17756 152.586 0.564304 153.157 0H115.663C119.697 0.197302 123.535 1.7608 126.518 4.42209C129.502 7.08358 131.444 10.677 132.012 14.5857C132.049 15.1253 132.049 15.6669 132.012 16.2065C131.967 20.329 130.796 24.3648 128.619 27.8965C126.663 31.0715 123.958 33.7355 120.738 35.6705ZM115.662 0H90.1248C90.4414 0.000747419 90.758 0.010437 91.0742 0.0290688C92.7228 0.283817 93.1625 2.50372 93.1625 3.85025C93.1625 5.19678 92.5764 7.48951 90.598 7.48951H88.8395C86.7465 7.66874 84.7134 8.27582 82.8678 9.27271C79.7688 10.6467 77.0263 12.7048 74.8487 15.2905C72.6712 17.8761 71.1161 20.9211 70.3018 24.1937C70.2835 24.4238 70.2835 24.655 70.3018 24.8852C70.3018 28.5244 74.9546 29.9074 77.9953 29.9074C81.9505 29.9074 86.6548 27.4332 89.3365 26.0227L89.3366 26.0227C90.4057 25.4603 91.1534 25.0671 91.404 25.0671C93.419 25.0671 95.0676 27.6874 95.0676 29.507C95.013 32.1645 91.1885 34.2473 86.841 35.6705H103.088C101.973 34.9951 100.95 34.1795 100.045 33.242C97.6088 30.4975 96.2431 27.0003 96.1885 23.3674C96.4911 18.7868 97.8067 14.3247 100.045 10.2892C101.579 7.44906 103.807 5.02353 106.532 3.22842C109.256 1.4334 112.393 0.324351 115.662 0ZM114.39 7.57556C117.746 7.97577 120.827 9.58911 123.027 12.0982V12.249C124.319 13.9005 124.987 15.9378 124.917 18.0155C124.969 19.8361 124.628 21.6471 123.914 23.3297C122.904 25.2127 121.437 26.8255 119.642 28.0277C117.848 29.2298 115.78 29.9849 113.618 30.2269C113.209 30.2812 112.794 30.2812 112.384 30.2269C106.446 29.0586 103.592 26.571 103.592 20.4277C103.551 18.6672 103.871 16.9164 104.534 15.2787C105.196 13.6411 106.187 12.1496 107.448 10.8922C109.224 8.93488 111.719 7.74269 114.39 7.57556ZM67.0034 34.7873C67.3286 35.1033 67.6725 35.3982 68.0329 35.6705H64.8181V31.7847C65.3681 32.9001 66.1065 33.916 67.0034 34.7873ZM64.8181 0V21.3462C66.0651 17.4026 68.1799 13.7591 71.0346 10.6921C75.5774 5.01483 81.4391 0.0290688 89.0959 0.0290688C89.4122 0.010437 89.7288 0.000747419 90.0454 0H64.8181ZM171.93 9.08119C173.739 8.70196 175.587 8.52074 177.438 8.54089V8.36083C178.814 8.36083 184.917 8.68493 185.289 10.522C185.289 11.0408 182.538 12.051 181.313 12.5008C181.092 12.5819 180.92 12.6448 180.824 12.6833C177.996 14.0881 175.093 15.2048 172.191 16.2854C171.815 15.034 171.639 13.7341 171.67 12.4311C171.664 11.3095 171.751 10.1893 171.93 9.08119Z" fill="white"/>
-    </Icon>
-  );
-};
-
-
-export default LogoWhiteIcon;
-
-
