+++ src/components/common/New/@Icons/LogoPrimary.tsx
@@ -1,0 +1,17 @@
+import { Icon, IconProps } from '@chakra-ui/react';
+
+const LogoPrimaryIcon = ({ ...props }: IconProps) => {
+  const w = props.w? props.w : "120px"
+  const h = props.h? props.h : "16px"
+  return (
+    <Icon
+      width={w} height={h} viewBox="0 0 120 16" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+      <path d="M0 0H6.02436V16.4834H0.111124L0 0Z" fill="#FF710B"/>
+      <path d="M7.53125 0H12.5357L19.0535 9.13404V0H24.0982V16.4834H19.0535L12.576 7.44539V16.4834H7.53125V0Z" fill="#FF710B"/>
+      <path d="M110.578 9.78771C113.391 9.08541 116.283 8.7298 119.187 8.7292C119.522 8.7166 119.793 8.57165 120 8.35463V16.3376H115.024C116.276 15.9025 117.51 15.4158 118.722 14.8786C118.926 14.7866 119.096 14.6359 119.21 14.4467C119.324 14.2575 119.376 14.0388 119.359 13.8201C119.359 13.0304 118.653 11.8039 117.74 11.8039C117.656 11.777 117.566 11.777 117.482 11.8039C115.313 12.8811 112.992 13.6398 110.595 14.0553C110.582 13.6803 110.562 13.3101 110.541 12.9411C110.508 12.3347 110.475 11.7313 110.475 11.115C110.458 10.6702 110.492 10.2249 110.578 9.78771ZM107.681 16.3376H98.6952C99.3658 15.9784 100.007 15.5645 100.612 15.099C102.386 13.7328 103.79 11.9637 104.702 9.946C104.985 9.31822 105.127 8.63856 105.118 7.95283C105.118 6.93935 104.72 6.11163 103.56 6.11163C103.396 6.11163 103.231 6.13275 103.067 6.15387C102.902 6.17499 102.738 6.19612 102.573 6.19612C100.721 6.36503 98.8693 6.51706 97.0172 6.51706C95.7826 6.51447 94.5499 6.41844 93.3305 6.22988C94.0797 5.34547 95.0006 4.61419 96.0402 4.07794C97.0798 3.54169 98.2177 3.21101 99.3885 3.10497H100.86C101.812 3.10497 102.833 2.98673 102.833 1.83811C102.836 1.32038 102.67 0.815187 102.359 0.395946C102.251 0.250315 102.127 0.117699 101.99 0H115.337C115.157 1.09654e-06 114.977 0.00307219 114.796 0.00921893C112.747 0.00921893 110.664 0.126852 108.615 0.294867C107.789 0.362073 107.444 1.7062 107.444 2.34465C107.417 2.83392 107.569 3.31672 107.875 3.70557V6.34342C107.634 6.34224 107.395 6.39055 107.174 6.48529C106.953 6.58004 106.755 6.71906 106.592 6.89333C106.43 7.06761 106.307 7.27323 106.231 7.49688C106.156 7.72053 106.129 7.95725 106.153 8.19159C106.14 8.57091 106.259 8.94342 106.49 9.24851C106.721 9.5536 107.052 9.77329 107.427 9.87175C107.367 10.9814 107.43 12.0969 107.493 13.208C107.538 13.9979 107.582 14.7855 107.582 15.5674C107.593 15.839 107.624 16.0979 107.681 16.3376ZM115.337 0C116.235 5.45979e-06 117.131 0.07612 118.016 0.227647C118.722 0.395662 118.894 1.33655 118.894 1.9078C118.894 2.47905 118.584 3.73915 117.947 3.73915C117.155 3.73915 116.372 3.63835 115.588 3.53754C114.805 3.43674 114.021 3.33594 113.229 3.33594C112.46 3.32745 111.7 3.49431 111.008 3.82319C110.918 4.07139 110.872 4.33269 110.871 4.59605C110.906 5.07148 110.906 5.54876 110.871 6.02419C111.551 5.97466 112.228 5.9149 112.904 5.85526C114.521 5.71258 116.131 5.57051 117.758 5.57051C118.024 5.57051 118.287 5.52852 118.55 5.48653C118.812 5.44453 119.075 5.40254 119.342 5.40254C119.616 5.40254 119.832 5.47502 120 5.59779V0H115.337ZM87.4847 16.3376H91.9194C91.74 15.9899 91.6342 15.631 91.6342 15.3174C91.6295 15.146 91.6601 14.9754 91.7242 14.8158C91.7882 14.6561 91.8844 14.5106 92.007 14.3878C92.1296 14.265 92.2762 14.1674 92.4381 14.1007C92.6 14.0341 92.774 13.9999 92.9497 13.9999C95.2252 13.903 97.4217 13.1572 99.2674 11.8547C99.9324 11.36 100.422 10.675 100.669 9.89533H98.627C96.723 9.89533 94.8017 10.0304 92.8977 10.1994H92.2573C91.6688 10.1994 89.8513 10.0642 89.5571 9.45614C89.328 8.95074 89.2101 8.40412 89.2109 7.85149C89.3445 7.00513 89.6488 6.19302 90.1064 5.46232C90.564 4.73163 91.1656 4.09687 91.8765 3.59483C93.3728 1.89243 95.3211 0.654498 97.4822 0H73.2935L73.1803 0.111217C73.1803 3.5199 73.0764 6.9286 72.9551 10.4566C72.9491 12.0843 72.3973 13.6647 71.3853 14.9527C70.9724 15.4782 70.4922 15.9431 69.9598 16.3376H75.6224C75.6438 15.8952 75.6915 15.4531 75.7392 15.0122L75.7392 15.0121C75.8029 14.4223 75.8664 13.8346 75.8664 13.2509C75.8664 12.2542 75.5464 11.9131 75.2644 11.6126C75.0423 11.3758 74.8438 11.1642 74.8438 10.6772C74.8817 9.93838 75.1898 9.23664 75.7129 8.69742C75.7129 8.37364 75.7345 8.05405 75.756 7.73653V7.73649V7.73644C75.7771 7.423 75.7982 7.11157 75.7982 6.80014C75.7957 5.97903 75.7501 5.15858 75.6618 4.34193C75.457 4.19275 75.2927 3.99772 75.1829 3.7736C75.0731 3.54947 75.021 3.30294 75.0312 3.05507C75.0511 2.4101 75.319 1.79553 75.7824 1.3317C76.2457 0.867864 76.8713 0.588132 77.5365 0.54732H78.1331C81.4991 0.367039 84.8668 0.872885 88.0181 2.03215C88.3583 2.19839 88.6411 2.45705 88.8318 2.77645C89.0225 3.09584 89.1128 3.46215 89.0918 3.83049C89.0918 5.77835 85.1087 8.05784 83.0877 9.21449L83.0876 9.21453L83.0856 9.21569C82.8992 9.32234 82.7296 9.41943 82.5813 9.50581C84.8907 10.2998 87.0244 11.5089 88.8703 13.0694C89.0982 13.3194 89.22 13.6436 89.2111 13.9769C89.1259 14.6863 88.4953 16.3361 87.4898 16.3361L87.4847 16.3376ZM87.2222 16.3376L87.2171 16.3361C86.7015 16.0213 86.2008 15.6922 85.7011 15.3638L85.701 15.3638C85.1269 14.9865 84.5541 14.6101 83.9619 14.2573C82.743 13.534 81.4794 12.8839 80.1783 12.3105C79.789 12.1249 79.3667 12.0129 78.9341 11.9805H78.8318C78.7072 13.1147 78.6786 14.257 78.7466 15.3957C78.7551 15.5495 78.7551 15.7037 78.7466 15.8576C78.702 16.0235 78.6415 16.1841 78.5664 16.3376H87.2222ZM55.2989 16.3376H62.2058C62.1083 16.1033 62.0541 15.8371 62.0541 15.5355C61.9963 14.6247 62.0157 13.691 62.035 12.765C62.0446 12.3049 62.0541 11.8466 62.0541 11.394C62.0541 9.80073 62.0758 8.2117 62.0974 6.62268L62.0975 6.62184L62.0975 6.62101C62.1191 5.03198 62.1408 3.44296 62.1408 1.84968C62.1408 0.963421 63.2673 0.281661 64.0991 0.281661C64.2534 0.265557 64.4093 0.283618 64.5555 0.334454C64.7017 0.385289 64.8345 0.467631 64.9442 0.575479C65.0539 0.683327 65.1376 0.81392 65.1893 0.957718C65.241 1.10152 65.2593 1.25487 65.243 1.40653V1.54291C65.243 2.82969 65.1606 4.10794 65.0783 5.38618C64.996 6.66443 64.9137 7.94268 64.9137 9.22946C64.867 10.3685 64.925 11.5094 65.087 12.6381C65.243 13.3369 65.5549 13.9335 66.3694 13.9335C66.7711 13.9014 67.1611 13.7851 67.5133 13.5926C70.0262 12.3995 70.0782 11.019 70.0782 8.47954C70.0782 6.7752 69.9742 5.07086 69.8702 3.36652C69.8702 2.95106 69.8316 2.52717 69.7928 2.10114C69.7536 1.67078 69.7142 1.23824 69.7142 0.810003C69.7142 0.540428 69.8876 0.258918 70.15 0H52.9738H41.2597C41.4107 0 41.5618 0.00444656 41.7126 0.0133354C42.4677 0.130013 42.6691 1.14676 42.6691 1.76349C42.6691 2.38021 42.4006 3.43032 41.4945 3.43032H40.6891C39.7305 3.5124 38.7993 3.79046 37.954 4.24705C36.5346 4.87634 35.2785 5.81901 34.2812 7.00326C33.2838 8.18751 32.5715 9.58215 32.1986 11.0811C32.1902 11.1865 32.1902 11.2924 32.1986 11.3978C32.1986 13.0646 34.3296 13.698 35.7223 13.698C37.5339 13.698 39.6885 12.5648 40.9168 11.9187L40.9168 11.9187L40.9168 11.9187C41.4065 11.6612 41.7489 11.4811 41.8637 11.4811C42.7866 11.4811 43.5417 12.6812 43.5417 13.5146C43.5167 14.7318 41.765 15.6857 39.7739 16.3376H47.2146C46.7042 16.0282 46.2354 15.6547 45.8209 15.2253C44.7052 13.9683 44.0797 12.3665 44.0547 10.7026C44.1932 8.6046 44.7958 6.56094 45.8209 4.71261C46.5235 3.41179 47.5441 2.30086 48.7921 1.47868C50.04 0.656497 51.4765 0.148529 52.9738 0C54.8216 0.0902751 56.5796 0.806394 57.946 2.0254C59.3125 3.2444 60.2022 4.89024 60.4624 6.68048C60.4793 6.92763 60.4793 7.17565 60.4624 7.4228C60.4416 9.31098 59.9051 11.1594 58.9084 12.777C58.0124 14.2312 56.7736 15.4513 55.2989 16.3376ZM30.6879 15.9331C30.8368 16.0778 30.9943 16.2128 31.1594 16.3376H29.6875V14.5588C29.9393 15.0693 30.2774 15.5342 30.6879 15.9331ZM29.6875 0V9.77532C30.2587 7.96966 31.2271 6.30147 32.5342 4.89713C34.6149 2.29688 37.2996 0.0133354 40.8066 0.0133354C40.9574 0.00444656 41.1085 0 41.2595 0H29.6875ZM52.391 3.46973C53.9282 3.65303 55.3395 4.39196 56.3472 5.54117V5.61023C56.9389 6.36663 57.2446 7.29975 57.2126 8.25135C57.2367 9.08523 57.0803 9.91469 56.7535 10.6853C56.2905 11.5478 55.6188 12.2865 54.7969 12.8371C53.9749 13.3877 53.0277 13.7335 52.0378 13.8444C51.8502 13.8692 51.6601 13.8692 51.4726 13.8444C48.7527 13.3092 47.4457 12.1699 47.4457 9.35616C47.4269 8.54983 47.5736 7.74795 47.8769 6.99788C48.1802 6.2478 48.6341 5.56471 49.2119 4.98881C50.025 4.09231 51.1678 3.54628 52.391 3.46973ZM78.7466 4.16042C79.5752 3.98673 80.4213 3.90373 81.269 3.91296V3.83049C81.8996 3.83049 84.6947 3.97893 84.8651 4.82034C84.8651 5.05797 83.605 5.52063 83.0438 5.72666L83.0437 5.72671L83.0432 5.7269C82.9423 5.76392 82.8641 5.79264 82.8199 5.81024C81.5247 6.45367 80.1953 6.96511 78.8659 7.46006C78.6937 6.88688 78.6132 6.29152 78.6273 5.69474C78.6246 5.18103 78.6645 4.66795 78.7466 4.16042Z" fill="#FF710B"/>
+    </Icon>
+  );
+};
+
+
+export default LogoPrimaryIcon;
