+++ src/components/common/New/@Icons/System/Cart.tsx
@@ -5,7 +5,7 @@ const CartIcon = ({ ...props }: IconProps) => {
   const h = props.h? props.h : "24px"
   return (
     <Icon
-      width={w} height={h} viewBox="0 0 24px 24px" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+      width={w} height={h} viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
       <path d="M17.9846 15.972H5.86455C5.77787 15.9677 5.69487 15.9362 5.62768 15.8822C5.56048 15.8283 5.51259 15.7546 5.49101 15.672L2.64114 6.77202C2.61949 6.70386 2.61256 6.632 2.6208 6.56106C2.62903 6.49011 2.65225 6.42164 2.68897 6.36002C2.72156 6.30293 2.76874 6.25519 2.82587 6.22152C2.88299 6.18784 2.94807 6.1694 3.01467 6.16802H17.9979C18.111 6.17672 18.216 6.22882 18.2903 6.31303C18.3646 6.39725 18.4021 6.50681 18.3948 6.61802C18.4021 6.72923 18.3646 6.83879 18.2903 6.92301C18.216 7.00722 18.111 7.05931 17.9979 7.06802H3.57956L6.14343 15.068H17.6793L20.4478 4.06802C20.4653 3.97892 20.5131 3.89831 20.5834 3.83944C20.6536 3.78056 20.7422 3.74691 20.8345 3.74402H22.6768C22.7899 3.75247 22.8951 3.80449 22.9695 3.88877C23.0438 3.97304 23.0813 4.08275 23.0737 4.19402C23.0813 4.30529 23.0438 4.415 22.9695 4.49927C22.8951 4.58354 22.7899 4.63557 22.6768 4.64402H21.1348L18.3633 15.644C18.3462 15.7323 18.2994 15.8124 18.2303 15.8714C18.1613 15.9304 18.074 15.9648 17.9826 15.969" fill="#1A1A1A"/>
       <path d="M17.9846 15.972H5.86455C5.77787 15.9677 5.69487 15.9362 5.62768 15.8822C5.56048 15.8283 5.51259 15.7546 5.49101 15.672L2.64114 6.77202C2.61949 6.70386 2.61256 6.632 2.6208 6.56106C2.62903 6.49011 2.65225 6.42164 2.68897 6.36002C2.72156 6.30293 2.76874 6.25519 2.82587 6.22152C2.88299 6.18784 2.94807 6.1694 3.01467 6.16802H17.9979C18.111 6.17672 18.216 6.22882 18.2903 6.31303C18.3646 6.39725 18.4021 6.50681 18.3948 6.61802C18.4021 6.72923 18.3646 6.83879 18.2903 6.92301C18.216 7.00722 18.111 7.05931 17.9979 7.06802H3.57956L6.14343 15.068H17.6793L20.4478 4.06802C20.4653 3.97892 20.5131 3.89831 20.5834 3.83944C20.6536 3.78056 20.7422 3.74691 20.8345 3.74402H22.6768C22.7899 3.75247 22.8951 3.80449 22.9695 3.88877C23.0438 3.97304 23.0813 4.08275 23.0737 4.19402C23.0813 4.30529 23.0438 4.415 22.9695 4.49927C22.8951 4.58354 22.7899 4.63557 22.6768 4.64402H21.1348L18.3633 15.644C18.3462 15.7323 18.2994 15.8124 18.2303 15.8714C18.1613 15.9304 18.074 15.9648 17.9826 15.969" stroke="#1A1A1A" stroke-width="0.7"/>
       <path d="M7.43204 21.482C7.01402 21.4892 6.60329 21.3739 6.25218 21.1509C5.90106 20.9279 5.62545 20.6073 5.46046 20.2299C5.29547 19.8524 5.24857 19.4353 5.32572 19.0316C5.40287 18.6278 5.60059 18.2558 5.89369 17.9629C6.18679 17.67 6.562 17.4694 6.97153 17.3867C7.38106 17.304 7.80638 17.343 8.19329 17.4986C8.58021 17.6543 8.91122 17.9195 9.14415 18.2606C9.37708 18.6018 9.50139 19.0033 9.50126 19.414C9.50559 19.9578 9.29012 20.4809 8.90215 20.8687C8.51418 21.2564 7.98544 21.477 7.43204 21.482ZM7.43204 18.245C7.19614 18.2411 6.96437 18.3062 6.76623 18.432C6.5681 18.5579 6.41256 18.7388 6.31943 18.9518C6.22629 19.1648 6.19977 19.4002 6.24323 19.628C6.28669 19.8559 6.39818 20.0659 6.56349 20.2313C6.72881 20.3967 6.94048 20.51 7.17155 20.5568C7.40263 20.6036 7.64266 20.5818 7.86109 20.4942C8.07952 20.4066 8.26647 20.2571 8.39815 20.0647C8.52982 19.8724 8.60026 19.6459 8.60049 19.414C8.60294 19.1069 8.4813 18.8114 8.26225 18.5922C8.04321 18.3731 7.74465 18.2482 7.43204 18.245Z" fill="#1A1A1A" stroke="#1A1A1A" stroke-width="0.7"/>
