+++ src/components/common/New/@Icons/System/Menu.tsx
@@ -5,7 +5,7 @@ const MenuIcon = ({ ...props }: IconProps) => {
   const h = props.h? props.h : "24px"
   return (
     <Icon
-      width={w} height={h} viewBox="0 0 24px 24px" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+      width={w} height={h} viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
       <path d="M21 5H3V7.5H21V5Z" fill="#1A1A1A"/>
       <path d="M21 16H3V18.5H21V16Z" fill="#1A1A1A"/>
     </Icon>
